# apps/backend/openapi/web-api.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "yaaos",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/tickets": {
      "get": {
        "description": "List tickets per the contract.\n\nReturns `{items, next_cursor}` instead of a bare array so the SPA can\ndrive Load-more pagination. `next_cursor` is null today — POC uses\nnaive limit pagination; opaque-cursor support lands when the result\nsets grow.",
        "operationId": "list__api_tickets_get",
        "parameters": [
          {
            "in": "query",
            "name": "repo_external_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Repo External Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "updated_desc",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "created_after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created After"
            }
          },
          {
            "in": "query",
            "name": "created_before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created Before"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Org-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Slug"
            }
          },
          {
            "in": "cookie",
            "name": "yaaos_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Yaaos Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List  Api Tickets Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List ",
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/tickets/dashboard": {
      "get": {
        "description": "Single-query Dashboard projection per E2a.3.\n\nResponse shape:\n`{stats: {in_flight, hitl_pending, completed_today, failed_today},\n   in_flight: [TicketRow ≤10],\n   needs_attention: [TicketRow ≤5]}`.\n\nAvoids the SPA making three `/api/tickets?status=…` calls in a tight\npolling loop. `t.status` is the 5-state vocab post-collapse\n(running / hitl / done / failed / cancelled); precise hitl/failed\ncounts depend on the workflow-state projection landing on every\ntransition.",
        "operationId": "dashboard_api_tickets_dashboard_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Org-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Slug"
            }
          },
          {
            "in": "cookie",
            "name": "yaaos_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Yaaos Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Dashboard Api Tickets Dashboard Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dashboard",
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/tickets/{ticket_id}": {
      "get": {
        "description": "Per-ticket detail with the enrichment: `stages[]` (projected from\n`workflow_executions`) + `builder` (the trigger identity).\n\nReturns the Ticket pydantic fields plus:\n- `stages: [{name, state, attempt_count, current_attempt, started_at,\n   completed_at, workflow_execution_id}]` — one entry per workflow run\n   on the ticket, newest first.\n- `builder: {kind, user_id?, display_name, avatar_url?}` — `kind=\"user\"`\n   when the ticket's PR has an `author_login`; `kind=\"system\"` when\n   yaaos triggered the run with no human attribution.",
        "operationId": "detail_api_tickets__ticket_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Org-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Slug"
            }
          },
          {
            "in": "cookie",
            "name": "yaaos_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Yaaos Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Detail Api Tickets  Ticket Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detail",
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/tickets/{ticket_id}/audit": {
      "get": {
        "description": "Aggregated timeline: ticket events + PR events for the ticket's PR.",
        "operationId": "audit_api_tickets__ticket_id__audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Org-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Slug"
            }
          },
          {
            "in": "cookie",
            "name": "yaaos_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Yaaos Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Audit Api Tickets  Ticket Id  Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit",
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/tickets/{ticket_id}/hitl/history": {
      "get": {
        "description": "List past HITL exchanges (prompt + response + timestamps) for the\nticket per E2a.4 HITL tab \"History\" subsection.\n\nJoins `pending_human_decisions` against the ticket's\n`workflow_executions` rows. Newest exchange first.",
        "operationId": "hitl_history_api_tickets__ticket_id__hitl_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Org-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Slug"
            }
          },
          {
            "in": "cookie",
            "name": "yaaos_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Yaaos Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Hitl History Api Tickets  Ticket Id  Hitl History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Hitl History",
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/tickets/{ticket_id}/hitl/respond": {
      "post": {
        "description": "Submit a HITL response. Resolves the open `PendingHumanDecisionRow`\nfor the ticket's most recent awaiting-human workflow execution and\nre-enqueues the routing step via `core.workflow.resume_hitl`.\n\nRequest body: opaque dict — passes through to the workflow engine's\n`resume_hitl(response=...)`. The SPA's HITL renderer shapes this\nper the prompt's discriminated-union schema (E2a.4).\n\nReturns `{stage, next_state}` where `next_state` is the workflow\nstate immediately after the resume.",
        "operationId": "hitl_respond_api_tickets__ticket_id__hitl_respond_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Org-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Slug"
            }
          },
          {
            "in": "cookie",
            "name": "yaaos_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Yaaos Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Response",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Hitl Respond Api Tickets  Ticket Id  Hitl Respond Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Hitl Respond",
        "tags": [
          "tickets"
        ]
      }
    },
...
  }
}